+++ source/tests/tf/test_deeppot_a.py
@@ -3,7 +3,6 @@ import os
 import shutil
 import unittest
 
-import dpdata
 import numpy as np
 from packaging.version import parse as parse_version
 
@@ -96,453 +95,7 @@ class TestModelMinorCompatability(unittest.TestCase):
         self.assertTrue("0.1000000" in str(context.exception))
 
 
-# TestDeepPotAPBC: moved to infer/test_models.py
-
-
-class TestDeepPotANoPBC(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        convert_pbtxt_to_pb(
-            str(infer_path / os.path.join("deeppot.pbtxt")), "deeppot.pb"
-        )
-        cls.dp = DeepPot("deeppot.pb")
-
-    def setUp(self):
-        self.coords = np.array(
-            [
-                12.83,
-                2.56,
-                2.18,
-                12.09,
-                2.87,
-                2.74,
-                00.25,
-                3.32,
-                1.68,
-                3.36,
-                3.00,
-                1.81,
-                3.51,
-                2.51,
-                2.60,
-                4.27,
-                3.22,
-                1.56,
-            ]
-        )
-        self.atype = [0, 1, 1, 0, 1, 1]
-        self.box = None
-        self.expected_e = np.array(
-            [
-                -9.255934839310273787e01,
-                -1.863253376736990106e02,
-                -1.857237299341402945e02,
-                -9.279308539717486326e01,
-                -1.863708105823244239e02,
-                -1.863635196514972563e02,
-            ]
-        )
-        self.expected_f = np.array(
-            [
-                -2.161037360255332107e00,
-                9.052994347015581589e-01,
-                1.635379623977007979e00,
-                2.161037360255332107e00,
-                -9.052994347015581589e-01,
-                -1.635379623977007979e00,
-                -1.167128117249453811e-02,
-                1.371975700096064992e-03,
-                -1.575265180249604477e-03,
-                6.226508593971802341e-01,
-                -1.816734122009256991e-01,
-                3.561766019664774907e-01,
-                -1.406075393906316626e-02,
-                3.789140061530929526e-01,
-                -6.018777878642909140e-01,
-                -5.969188242856223736e-01,
-                -1.986125696522633155e-01,
-                2.472764510780630642e-01,
-            ]
-        )
-        self.expected_v = np.array(
-            [
-                -7.042445481792056761e-01,
-                2.950213647777754078e-01,
-                5.329418202437231633e-01,
-                2.950213647777752968e-01,
-                -1.235900311906896754e-01,
-                -2.232594111831812944e-01,
-                5.329418202437232743e-01,
-                -2.232594111831813499e-01,
-                -4.033073234276823849e-01,
-                -8.949230984097404917e-01,
-                3.749002169013777030e-01,
-                6.772391014992630298e-01,
-                3.749002169013777586e-01,
-                -1.570527935667933583e-01,
-                -2.837082722496912512e-01,
-                6.772391014992631408e-01,
-                -2.837082722496912512e-01,
-                -5.125052659994422388e-01,
-                4.858210330291591605e-02,
-                -6.902596153269104431e-03,
-                6.682612642430500391e-03,
-                -5.612247004554610057e-03,
-                9.767795567660207592e-04,
-                -9.773758942738038254e-04,
-                5.638322117219018645e-03,
-                -9.483806049779926932e-04,
-                8.493873281881353637e-04,
-                -2.941738570564985666e-01,
-                -4.482529909499673171e-02,
-                4.091569840186781021e-02,
-                -4.509020615859140463e-02,
-                -1.013919988807244071e-01,
-                1.551440772665269030e-01,
-                4.181857726606644232e-02,
-                1.547200233064863484e-01,
-                -2.398213304685777592e-01,
-                -3.218625798524068354e-02,
-                -1.012438450438508421e-02,
-                1.271639330380921855e-02,
-                3.072814938490859779e-03,
-                -9.556241797915024372e-02,
-                1.512251983492413077e-01,
-                -8.277872384009607454e-03,
-                1.505412040827929787e-01,
-                -2.386150620881526407e-01,
-                -2.312295470054945568e-01,
-                -6.631490213524345034e-02,
-                7.932427266386249398e-02,
-                -8.053754366323923053e-02,
-                -3.294595881137418747e-02,
-                4.342495071150231922e-02,
-                1.004599500126941436e-01,
-                4.450400364869536163e-02,
-                -5.951077548033092968e-02,
-            ]
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        os.remove("deeppot.pb")
-        cls.dp = None
-
-    def test_1frame(self):
-        ee, ff, vv = self.dp.eval(self.coords, self.box, self.atype, atomic=False)
-        # check shape of the returns
-        nframes = 1
-        natoms = len(self.atype)
-        self.assertEqual(ee.shape, (nframes, 1))
-        self.assertEqual(ff.shape, (nframes, natoms, 3))
-        self.assertEqual(vv.shape, (nframes, 9))
-        # check values
-        np.testing.assert_almost_equal(
-            ff.ravel(), self.expected_f.ravel(), default_places
-        )
-        expected_se = np.sum(self.expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(ee.ravel(), expected_se.ravel(), default_places)
-        expected_sv = np.sum(self.expected_v.reshape([nframes, -1, 9]), axis=1)
-        np.testing.assert_almost_equal(vv.ravel(), expected_sv.ravel(), default_places)
-
-    def test_1frame_atm(self):
-        ee, ff, vv, ae, av = self.dp.eval(
-            self.coords, self.box, self.atype, atomic=True
-        )
-        # check shape of the returns
-        nframes = 1
-        natoms = len(self.atype)
-        self.assertEqual(ee.shape, (nframes, 1))
-        self.assertEqual(ff.shape, (nframes, natoms, 3))
-        self.assertEqual(vv.shape, (nframes, 9))
-        self.assertEqual(ae.shape, (nframes, natoms, 1))
-        self.assertEqual(av.shape, (nframes, natoms, 9))
-        # check values
-        np.testing.assert_almost_equal(
-            ff.ravel(), self.expected_f.ravel(), default_places
-        )
-        np.testing.assert_almost_equal(
-            ae.ravel(), self.expected_e.ravel(), default_places
-        )
-        np.testing.assert_almost_equal(
-            av.ravel(), self.expected_v.ravel(), default_places
-        )
-        expected_se = np.sum(self.expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(ee.ravel(), expected_se.ravel(), default_places)
-        expected_sv = np.sum(self.expected_v.reshape([nframes, -1, 9]), axis=1)
-        np.testing.assert_almost_equal(vv.ravel(), expected_sv.ravel(), default_places)
-
-    def test_2frame_atm(self):
-        coords2 = np.concatenate((self.coords, self.coords))
-        ee, ff, vv, ae, av = self.dp.eval(coords2, self.box, self.atype, atomic=True)
-        # check shape of the returns
-        nframes = 2
-        natoms = len(self.atype)
-        self.assertEqual(ee.shape, (nframes, 1))
-        self.assertEqual(ff.shape, (nframes, natoms, 3))
-        self.assertEqual(vv.shape, (nframes, 9))
-        self.assertEqual(ae.shape, (nframes, natoms, 1))
-        self.assertEqual(av.shape, (nframes, natoms, 9))
-        # check values
-        expected_f = np.concatenate((self.expected_f, self.expected_f), axis=0)
-        expected_e = np.concatenate((self.expected_e, self.expected_e), axis=0)
-        expected_v = np.concatenate((self.expected_v, self.expected_v), axis=0)
-        np.testing.assert_almost_equal(ff.ravel(), expected_f.ravel(), default_places)
-        np.testing.assert_almost_equal(ae.ravel(), expected_e.ravel(), default_places)
-        np.testing.assert_almost_equal(av.ravel(), expected_v.ravel(), default_places)
-        expected_se = np.sum(expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(ee.ravel(), expected_se.ravel(), default_places)
-        expected_sv = np.sum(expected_v.reshape([nframes, -1, 9]), axis=1)
-        np.testing.assert_almost_equal(vv.ravel(), expected_sv.ravel(), default_places)
-
-    def test_dpdata_driver(self):
-        nframes = 1
-        system = dpdata.System(
-            data={
-                "coords": self.coords.reshape((nframes, -1, 3)),
-                "cells": np.zeros((nframes, 3, 3)),
-                "atom_types": np.array(self.atype),
-                "orig": np.zeros((3,)),
-                "atom_names": ["O", "H"],
-                "atom_numbs": [2, 4],
-                "nopbc": True,
-            }
-        )
-        system_predicted = system.predict(self.dp, driver="dp")
-        np.testing.assert_almost_equal(
-            system_predicted["forces"].ravel(), self.expected_f.ravel(), default_places
-        )
-        expected_se = np.sum(self.expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(
-            system_predicted["energies"].ravel(), expected_se.ravel(), default_places
-        )
-        expected_sv = np.sum(self.expected_v.reshape([nframes, -1, 9]), axis=1)
-        np.testing.assert_almost_equal(
-            system_predicted["virials"].ravel(), expected_sv.ravel(), default_places
-        )
-
-
-class TestDeepPotALargeBoxNoPBC(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        convert_pbtxt_to_pb(
-            str(infer_path / os.path.join("deeppot.pbtxt")), "deeppot.pb"
-        )
-        cls.dp = DeepPot("deeppot.pb")
-
-    def setUp(self):
-        self.coords = np.array(
-            [
-                12.83,
-                2.56,
-                2.18,
-                12.09,
-                2.87,
-                2.74,
-                00.25,
-                3.32,
-                1.68,
-                3.36,
-                3.00,
-                1.81,
-                3.51,
-                2.51,
-                2.60,
-                4.27,
-                3.22,
-                1.56,
-            ]
-        )
-        self.atype = [0, 1, 1, 0, 1, 1]
-        self.box = np.array([19.0, 0.0, 0.0, 0.0, 13.0, 0.0, 0.0, 0.0, 13.0])
-        self.expected_e = np.array(
-            [
-                -9.255934839310273787e01,
-                -1.863253376736990106e02,
-                -1.857237299341402945e02,
-                -9.279308539717486326e01,
-                -1.863708105823244239e02,
-                -1.863635196514972563e02,
-            ]
-        )
-        self.expected_f = np.array(
-            [
-                -2.161037360255332107e00,
-                9.052994347015581589e-01,
-                1.635379623977007979e00,
-                2.161037360255332107e00,
-                -9.052994347015581589e-01,
-                -1.635379623977007979e00,
-                -1.167128117249453811e-02,
-                1.371975700096064992e-03,
-                -1.575265180249604477e-03,
-                6.226508593971802341e-01,
-                -1.816734122009256991e-01,
-                3.561766019664774907e-01,
-                -1.406075393906316626e-02,
-                3.789140061530929526e-01,
-                -6.018777878642909140e-01,
-                -5.969188242856223736e-01,
-                -1.986125696522633155e-01,
-                2.472764510780630642e-01,
-            ]
-        )
-        self.expected_v = np.array(
-            [
-                -7.042445481792056761e-01,
-                2.950213647777754078e-01,
-                5.329418202437231633e-01,
-                2.950213647777752968e-01,
-                -1.235900311906896754e-01,
-                -2.232594111831812944e-01,
-                5.329418202437232743e-01,
-                -2.232594111831813499e-01,
-                -4.033073234276823849e-01,
-                -8.949230984097404917e-01,
-                3.749002169013777030e-01,
-                6.772391014992630298e-01,
-                3.749002169013777586e-01,
-                -1.570527935667933583e-01,
-                -2.837082722496912512e-01,
-                6.772391014992631408e-01,
-                -2.837082722496912512e-01,
-                -5.125052659994422388e-01,
-                4.858210330291591605e-02,
-                -6.902596153269104431e-03,
-                6.682612642430500391e-03,
-                -5.612247004554610057e-03,
-                9.767795567660207592e-04,
-                -9.773758942738038254e-04,
-                5.638322117219018645e-03,
-                -9.483806049779926932e-04,
-                8.493873281881353637e-04,
-                -2.941738570564985666e-01,
-                -4.482529909499673171e-02,
-                4.091569840186781021e-02,
-                -4.509020615859140463e-02,
-                -1.013919988807244071e-01,
-                1.551440772665269030e-01,
-                4.181857726606644232e-02,
-                1.547200233064863484e-01,
-                -2.398213304685777592e-01,
-                -3.218625798524068354e-02,
-                -1.012438450438508421e-02,
-                1.271639330380921855e-02,
-                3.072814938490859779e-03,
-                -9.556241797915024372e-02,
-                1.512251983492413077e-01,
-                -8.277872384009607454e-03,
-                1.505412040827929787e-01,
-                -2.386150620881526407e-01,
-                -2.312295470054945568e-01,
-                -6.631490213524345034e-02,
-                7.932427266386249398e-02,
-                -8.053754366323923053e-02,
-                -3.294595881137418747e-02,
-                4.342495071150231922e-02,
-                1.004599500126941436e-01,
-                4.450400364869536163e-02,
-                -5.951077548033092968e-02,
-            ]
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        os.remove("deeppot.pb")
-        cls.dp = None
-
-    def test_1frame(self):
-        ee, ff, vv = self.dp.eval(self.coords, self.box, self.atype, atomic=False)
-        # check shape of the returns
-        nframes = 1
-        natoms = len(self.atype)
-        self.assertEqual(ee.shape, (nframes, 1))
-        self.assertEqual(ff.shape, (nframes, natoms, 3))
-        self.assertEqual(vv.shape, (nframes, 9))
-        # check values
-        np.testing.assert_almost_equal(
-            ff.ravel(), self.expected_f.ravel(), default_places
-        )
-        expected_se = np.sum(self.expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(ee.ravel(), expected_se.ravel(), default_places)
-        expected_sv = np.sum(self.expected_v.reshape([nframes, -1, 9]), axis=1)
-        np.testing.assert_almost_equal(vv.ravel(), expected_sv.ravel(), default_places)
-
-    def test_1frame_atm(self):
-        ee, ff, vv, ae, av = self.dp.eval(
-            self.coords, self.box, self.atype, atomic=True
-        )
-        # check shape of the returns
-        nframes = 1
-        natoms = len(self.atype)
-        self.assertEqual(ee.shape, (nframes, 1))
-        self.assertEqual(ff.shape, (nframes, natoms, 3))
-        self.assertEqual(vv.shape, (nframes, 9))
-        self.assertEqual(ae.shape, (nframes, natoms, 1))
-        self.assertEqual(av.shape, (nframes, natoms, 9))
-        # check values
-        np.testing.assert_almost_equal(
-            ff.ravel(), self.expected_f.ravel(), default_places
-        )
-        np.testing.assert_almost_equal(
-            ae.ravel(), self.expected_e.ravel(), default_places
-        )
-        np.testing.assert_almost_equal(
-            av.ravel(), self.expected_v.ravel(), default_places
-        )
-        expected_se = np.sum(self.expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(ee.ravel(), expected_se.ravel(), default_places)
-        expected_sv = np.sum(self.expected_v.reshape([nframes, -1, 9]), axis=1)
-        np.testing.assert_almost_equal(vv.ravel(), expected_sv.ravel(), default_places)
-
-    def test_ase(self):
-        from ase import (
-            Atoms,
-        )
-
-        from deepmd.tf.calculator import (
-            DP,
-        )
-
-        water = Atoms(
-            "OHHOHH",
-            positions=self.coords.reshape((-1, 3)),
-            cell=self.box.reshape((3, 3)),
-            calculator=DP("deeppot.pb"),
-        )
-        ee = water.get_potential_energy()
-        ff = water.get_forces()
-        nframes = 1
-        np.testing.assert_almost_equal(
-            ff.ravel(), self.expected_f.ravel(), default_places
-        )
-        expected_se = np.sum(self.expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(ee.ravel(), expected_se.ravel(), default_places)
-
-    def test_dpdata_driver(self):
-        nframes = 1
-        system = dpdata.System(
-            data={
-                "coords": self.coords.reshape((nframes, -1, 3)),
-                "cells": self.box.reshape((nframes, 3, 3)),
-                "atom_types": np.array(self.atype),
-                "orig": np.zeros((3,)),
-                "atom_names": ["O", "H"],
-                "atom_numbs": [2, 4],
-            }
-        )
-        system_predicted = system.predict("deeppot.pb", driver="dp")
-        np.testing.assert_almost_equal(
-            system_predicted["forces"].ravel(), self.expected_f.ravel(), default_places
-        )
-        expected_se = np.sum(self.expected_e.reshape([nframes, -1]), axis=1)
-        np.testing.assert_almost_equal(
-            system_predicted["energies"].ravel(), expected_se.ravel(), default_places
-        )
-        expected_sv = np.sum(self.expected_v.reshape([nframes, -1, 9]), axis=1)
-        np.testing.assert_almost_equal(
-            system_predicted["virials"].ravel(), expected_sv.ravel(), default_places
-        )
+# TestDeepPotAPBC, TestDeepPotANoPBC, TestDeepPotALargeBoxNoPBC: moved to infer/test_models.py
 
 
 class TestModelConvert(unittest.TestCase):
